fixing socket
Firefox will trigger `write error: websocket: close sent` (chromium seems to be unable to trigger this event, even with thousands of refreshes sent) by pressing refresh several times on a page with a load time of (network latency included) >250ms 
This causes the `unregister <- connection` to fire, however this breaks the socket entirely for ALL current connections and all future connections will connect to the websocket but the websocket is down and will not transfer / receive messages. (Messages can be sent to the server ACCORDING TO THE BROWSER but without error but they're thrown out by the server without notice of any kind).

No idea why it happens like that, but it does. Posted fix works fine.