feat(recipe/tableflip): Add Graceful shutdown to ensure continuous uptime
In the original Cloudflare Tableflip http example, they use a graceful shutdown method to ensure a continuous service uptime. In contrast, with the original fiber recipe, if you send the upgrade signal to the fiber process, current connections will die out (you can test with `ab -c 50 -n 100000 http://127.0.0.1:8080/version` and then send an upgrade signal while the test is running and see how the service dies). This commit fixes this issue with the original recipe and now actually allows for a continuous service uptime.